refact: rendering

+++ src/ui/file_picker.rs
@@ -54,7 +54,7 @@ impl Component for FilePicker {
         self.picker.draw(buffer)
     }
 
-    fn cursor_position(&self) -> Option<(u16, u16)> {
+    fn cursor_position(&self) -> Option<(usize, usize)> {
         self.picker.cursor_position()
     }
 }
